second pass of reverting config move
