+++ internal/scheduler/queue_scheduler_test.go
@@ -423,16 +423,16 @@ func TestQueueScheduler(t *testing.T) {
 			Nodes:            testfixtures.N32CpuNodes(3, testfixtures.TestPriorities),
 			Jobs: armadaslices.Concatenate(
 				testfixtures.WithAnnotationsJobs(map[string]string{
-					apiconfiguration.GangIdAnnotation:                 "my-gang",
-					apiconfiguration.GangCardinalityAnnotation:        "2",
-					apiconfiguration.GangMinimumCardinalityAnnotation: "1",
+					configuration.GangIdAnnotation:                 "my-gang",
+					configuration.GangCardinalityAnnotation:        "2",
+					configuration.GangMinimumCardinalityAnnotation: "1",
 				},
 					testfixtures.N32Cpu256GiJobs("A", testfixtures.PriorityClass0, 1)),
 				testfixtures.N1Cpu4GiJobs("A", testfixtures.PriorityClass0, 1),
 				testfixtures.WithAnnotationsJobs(map[string]string{
-					apiconfiguration.GangIdAnnotation:                 "my-gang",
-					apiconfiguration.GangCardinalityAnnotation:        "2",
-					apiconfiguration.GangMinimumCardinalityAnnotation: "1",
+					configuration.GangIdAnnotation:                 "my-gang",
+					configuration.GangCardinalityAnnotation:        "2",
+					configuration.GangMinimumCardinalityAnnotation: "1",
 				},
 					testfixtures.N32Cpu256GiJobs("A", testfixtures.PriorityClass0, 1)),
 			),
@@ -451,16 +451,16 @@ func TestQueueScheduler(t *testing.T) {
 			Nodes:            testfixtures.N32CpuNodes(2, testfixtures.TestPriorities),
 			Jobs: armadaslices.Concatenate(
 				testfixtures.WithAnnotationsJobs(map[string]string{
-					apiconfiguration.GangIdAnnotation:                 "my-gang",
-					apiconfiguration.GangCardinalityAnnotation:        "2",
-					apiconfiguration.GangMinimumCardinalityAnnotation: "2",
+					configuration.GangIdAnnotation:                 "my-gang",
+					configuration.GangCardinalityAnnotation:        "2",
+					configuration.GangMinimumCardinalityAnnotation: "2",
 				},
 					testfixtures.N32Cpu256GiJobs("A", testfixtures.PriorityClass0, 1)),
 				testfixtures.N1Cpu4GiJobs("A", testfixtures.PriorityClass0, 1),
 				testfixtures.WithAnnotationsJobs(map[string]string{
-					apiconfiguration.GangIdAnnotation:                 "my-gang",
-					apiconfiguration.GangCardinalityAnnotation:        "2",
-					apiconfiguration.GangMinimumCardinalityAnnotation: "2",
+					configuration.GangIdAnnotation:                 "my-gang",
+					configuration.GangCardinalityAnnotation:        "2",
+					configuration.GangMinimumCardinalityAnnotation: "2",
 				},
 					testfixtures.N32Cpu256GiJobs("A", testfixtures.PriorityClass0, 1)),
 			),
