+++ internal/scheduler/scheduling_algo_test.go
@@ -480,7 +480,7 @@ func TestSchedule(t *testing.T) {
 			}
 			// Sanity check: we've set `GangNumJobsScheduledAnnotation` for all scheduled jobs.
 			for _, job := range scheduledJobs {
-				assert.Contains(t, schedulerResult.AdditionalAnnotationsByJobId[job.Id()], apiconfiguration.GangNumJobsScheduledAnnotation)
+				assert.Contains(t, schedulerResult.AdditionalAnnotationsByJobId[job.Id()], configuration.GangNumJobsScheduledAnnotation)
 			}
 
 			// Check that we failed the correct number of excess jobs when a gang schedules >= minimum cardinality
